rename block to bockchain and merge network command with it
