+++ wallet-cli/src/command/mod.rs
@@ -1,9 +1,7 @@
 mod blockchain;
-mod block;
 mod wallet;
 
 pub use self::blockchain::{*};
-pub use self::block::{*};
 pub use self::wallet::{*}; 
 
 use clap::{ArgMatches, App, SubCommand};
